separate shortcuts and methods
