+++ apps/testapp/src/components/RpcMethods/RpcMethodCard.tsx
@@ -22,7 +22,8 @@ import React, { useCallback } from 'react';
 import { useForm } from 'react-hook-form';
 
 import { useCBWSDK } from '../../context/CBWSDKProvider';
-import { ADDR_TO_FILL, verifySignMsg } from './methods/signMessageMethods';
+import { verifySignMsg } from './method/signMessageMethods';
+import { ADDR_TO_FILL } from './shortcut/const';
 
 type ResponseType = string;
