+++ apps/testapp/src/components/RpcMethods/shortcut/const.ts
@@ -1,0 +1,3 @@
+export const EXAMPLE_MESSAGE = `Example Message`;
+
+export const ADDR_TO_FILL = 'ADDR_TO_FILL';
